Wrap map fragment in multiplatform composable
