+++ app-android/src/main/java/org/pointyware/xyz/android/MainActivity.kt
@@ -7,13 +7,10 @@ package org.pointyware.xyz.android
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.fragment.app.FragmentActivity
-import androidx.fragment.compose.AndroidFragment
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.MarkerOptions
 import org.koin.mp.KoinPlatform.getKoin
+import org.pointyware.xyz.shared.XyzApp
 import org.pointyware.xyz.shared.di.AppDependencies
 
 class MainActivity : FragmentActivity() {
@@ -25,19 +22,10 @@ class MainActivity : FragmentActivity() {
         val appDependencies = koin.get<AppDependencies>()
 
         setContent {
-//            XyzApp(
-//                dependencies = appDependencies,
-//                isDarkTheme = isSystemInDarkTheme()
-//            )
-
-            AndroidFragment<SupportMapFragment> { mapFragment ->
-                mapFragment.getMapAsync { googleMap ->
-                    // Add a marker at the turn-off point of I-35 - 36°06'56.0"N 97°20'42.1"W
-                    val sydney = LatLng(36.11556, -97.34503)
-                    googleMap.addMarker(MarkerOptions().position(sydney).title("Marker in Atlantic Ocean"))
-                    googleMap.moveCamera(CameraUpdateFactory.newLatLngZoom(sydney, 11f))
-                }
-            }
+            XyzApp(
+                dependencies = appDependencies,
+                isDarkTheme = isSystemInDarkTheme()
+            )
         }
     }
 }
