+++ core/ui/src/commonMain/kotlin/maps/MapView.kt
@@ -1,0 +1,6 @@
+package org.pointyware.xyz.core.ui.maps
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun MapView()
