+++ core/ui/src/jvmMain/kotlin/maps/MapView.jvm.kt
@@ -1,0 +1,9 @@
+package org.pointyware.xyz.core.ui.maps
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun MapView() {
+    Text("MapView Under Construction on JVM")
+}
